requirements.txt
PyQt5==5.15.10
pandas==2.2.0
openpyxl==3.1.2
DateTime==5.5
mysql-connector-python==8.4.0